% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{build_mrf}
\alias{build_mrf}
\title{Train Multivariate Random Forest}
\usage{
build_mrf(trainX, trainY, n_tree, m_feature, min_leaf)
}
\arguments{
\item{trainX}{The design matrix for predictions. Can be made with create_X.}

\item{trainY}{The value of the response variables.}

\item{n_tree}{Number of trees in the forest, which must be positive integer.}

\item{m_feature}{Number of randomly selected features considered for a split in each regression tree node, which must be positive integer and less than N (number of input features)}

\item{min_leaf}{Minimum number of samples in the leaf node. If a node has less than or equal to min_leaf samples, then there will be no splitting in that node and this node will be considered as a leaf node. Valid input is positive integer, which is less than or equal to M (number of training samples)}
}
\value{
A named list including the trained model and it's predictions.
}
\description{
Similar to the build_forest_predict function from MultivariateRandomForest. However,
this function will save the model for future use.
}
