% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_D}
\alias{calc_D}
\title{Calculate D matrix}
\usage{
calc_D(tau, mu_tau, sigma_tau_inv)
}
\arguments{
\item{tau}{A vector representing covariance.}

\item{mu_tau}{The mean of tau. Estimated from the training data.}

\item{sigma_tau_inv}{The inverse covariance matrix of tau. Estimated from the training data.}
}
\value{
A vector D which is used to calculate the plotting statistic.
}
\description{
A helper function to calculate the D vector given tau,
mu_tau, and sigma_tau_inv. This is the mahalanobis distance for tau.
}
