% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PStat.R
\name{calc_PStat}
\alias{calc_PStat}
\title{Calculate plotting statistic}
\usage{
calc_PStat(S, sigma_tau_inv)
}
\arguments{
\item{S}{A matrix which tracks the MCUSUM through observations}

\item{sigma_tau_inv}{The inverse covariance matrix of tau. Estimated from the training data.}
}
\value{
The plotting statistic which determines if the process is considered in-control or out-of-control at a given observation.
}
\description{
A helper function to calculate the plotting statistic given
S and sigma_tau_inv.
}
