% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tau.R
\name{calc_tau}
\alias{calc_tau}
\title{Create tau vector}
\usage{
calc_tau(residuals)
}
\arguments{
\item{residuals}{The errors from one of our prediction methods on a Y matrix.}
}
\value{
The tau vector to be used for calculating the S vector.
}
\description{
A helper function to create the tau vector given a
vector of residuals from one of our methods.
}
