% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Y.R
\name{create_Y}
\alias{create_Y}
\title{Create Prediction Matrix}
\usage{
create_Y(data, lags)
}
\arguments{
\item{data}{A dataframe or matrix of observations.}

\item{lags}{The number of lags of each variable, the first row of this matrix will be row lag + 1 of the data.}
}
\value{
A prediction matrix which can be used for training/testing the machine learning MCUSUM methods.
}
\description{
A helper function to create a prediction matrix for
use when training the machine learning MCUSUM methods.
Requires choosing the number of lags so that the prediction
and design matrices line up.
}
