% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{gen_dat_lin}
\alias{gen_dat_lin}
\alias{gen_dat_nlr}
\alias{gen_dat_ltm}
\title{Generate Sample Data}
\usage{
gen_dat_lin(fault, n_ic = 1500, n_oc = 500, phi = 0.8)

gen_dat_nlr(fault, n_ic = 1500, n_oc = 500, phi = 0.8)

gen_dat_ltm(fault, n_ic = 1500, n_oc = 500, phi = 0.8)
}
\arguments{
\item{fault}{A marker for which fault is introduced, "f1", "f2", or "f3".}

\item{n_ic}{The number of observations which are in-control.}

\item{n_oc}{The number of observations which are out-of-control, i.e. faulty.}

\item{phi}{The strength of the autocorrelation in the base time series}
}
\value{
A dataframe of sample data for use in the simulation study.
}
\description{
Generates a three dimensional multivariate time series for use in the
simulation study. Functions exists for linearly related data, non-linearly
related data, and data with both long term dependencies and non-linear
relationships. Further, one of three faults can be specified via "f1", "f2",
or "f3".
}
