% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_first_fault}
\alias{get_first_fault}
\title{Calculate Run Length}
\usage{
get_first_fault(pstat, h)
}
\arguments{
\item{pstat}{The plotting statistic from a fault detection method.}

\item{h}{The upper in-control limit for this plotting statistic.}
}
\value{
Returns the run length.
}
\description{
Calculates the run length for a single instance of a method given a
plotting statistic and an upper limit by returning the index of the
first flagged observation.
}
