% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_python.R
\name{load_python}
\alias{load_python}
\title{Find python conda location}
\usage{
load_python(path_conda, path_py)
}
\arguments{
\item{path_conda}{A path for use_condaenv. This conda enviornment requires numpy and tensorflow.}

\item{path_py}{A path for gru_functions.py. This file contains to necessary python code to run the neural networks.}
}
\description{
Uses reticulate to load in the appropriate conda enviornment.
Will then use source_python to load in python source code for
using the GRU.
}
