% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_fd}
\alias{predict_fd}
\alias{predict_ml}
\title{Apply Method}
\usage{
predict_ml(model, new_data)
}
\arguments{
\item{model}{Output from the train_gruMCUSUM function.}

\item{new_data}{A multivariate time series in dataframe or matrix form.}
}
\value{
A named list including the plotting statistic and residuals.
}
\description{
Predicts values for new_data using the a trained GRU or MRF method. Also works
for the resMCUSUM method from Bodnar et al (2017) which is based on a VARMA(1, 1)
model. Returns the calculated residuals and the plotting statistic.
}
