% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gruMCUSUM.R
\name{predict_gruMCUSUM}
\alias{predict_gruMCUSUM}
\title{Predicts using gruMCUSUM method}
\usage{
predict_gruMCUSUM(model, new_data)
}
\arguments{
\item{model}{Output from the train_gruMCUSUM function.}

\item{new_data}{A multivariate time series in dataframe or matrix form.}
}
\value{
A named list including the plotting statistic and residuals.
}
\description{
Predicts the gruMCUSUM method.
}
