% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_mrf}
\alias{predict_mrf}
\title{Predict Multivariate Random Forest}
\usage{
predict_mrf(model, testX)
}
\arguments{
\item{model}{Output from build_mrf.}

\item{testX}{The design matrix for predictions. Can be made with create_X.}
}
\value{
A prediction matrix
}
\description{
Similar to the build_forest_predict function from MultivariateRandomForest. However,
this function takes in output from build_mrf and makes a new prediction.
}
