% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mrfMCUSUM.R
\name{predict_mrfMCUSUM}
\alias{predict_mrfMCUSUM}
\title{Predicts using mrfMCUSUM method}
\usage{
predict_mrfMCUSUM(model, new_data)
}
\arguments{
\item{model}{Output from the train_mrfMCUSUM function.}

\item{new_data}{A multivariate time series in dataframe or matrix form.}
}
\value{
A named list including the plotting statistic and residuals.
}
\description{
Predicts the mrfMCUSUM method.
}
