% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_t.R
\name{scale_t}
\alias{scale_t}
\title{Scale a TS}
\usage{
scale_t(ts, a = 0.01, b = 2)
}
\arguments{
\item{ts}{A univariate time series.}

\item{a}{Lower bound of time series, default is 0.01.}

\item{b}{Upper bound of time series, default is 2.}
}
\value{
Returns the scaled time series.
}
\description{
Scale a univariate time series between lower bound a and upper bound b.
Used for generating t for a simulation study.
}
