% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\alias{train_gruMCUSUM}
\alias{train_mrfMCUSUM}
\alias{train_varmaMCUSUM}
\alias{train_varmaMEWMA}
\title{Train methods}
\usage{
train_gruMCUSUM(data, lags = 1, k = 0.9)

train_mrfMCUSUM(data, lags = 1, k = 0.9)

train_varmaMCUSUM(data, k = 0.9)

train_varmaMEWMA(data, r = 0.3)
}
\arguments{
\item{data}{A multivariate time series in dataframe or matrix form.}

\item{lags}{The number of lags of each variable to be included in the design matrix.}

\item{k}{A tuning parameter for the MCUSUM, large k results in shorter memory.}

\item{r}{A tuning parameter for MEWMA, large r results in shorter memory.}
}
\value{
A named list including the plotting statistic, trained model, residuals, and constants.
}
\description{
Train the mrfMCUSUM, gruMCUSUM method. Also works for the resMCUSUM method
from Bodnar et al. (2017), which uses a VARMA(1, 1).
}
