% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_fd}
\alias{train_fd}
\title{Train methods}
\usage{
train_fd(
  data,
  method = "gruMEWMA",
  data_exog = NULL,
  lags = 1,
  k = 1.1,
  r = 0.3,
  center_scale = TRUE
)
}
\arguments{
\item{data}{A multivariate time series in dataframe or matrix form.}

\item{method}{An indicator of which model and fault detection method to use.
Options include gruMEWMA, mrfMCUSUM, varMEWMA, or htsquare.}

\item{data_exog}{Any exogenous variables to be considered in model training.
Must be a dataframe or matrix with the same number of rows as \code{data}.}

\item{lags}{The number of lags of each variable to be included in the design matrix.}

\item{k}{A tuning parameter for the MCUSUM, large k results in shorter memory.}

\item{r}{A tuning parameter for MEWMA, large r results in shorter memory.}

\item{center_scale}{A logical, whether or not to center and scale data before modeling.}
}
\value{
A named list including the plotting statistic, trained model, residuals, and constants.
}
\description{
Trains and applies a model based fault detection method for detecting the
changes in a multivariate time series. Can train a GRU, MRF, or VAR
model to pair with either an MCUSUM or MEWMA control chart. Can also apply
a centered and scaled Hotelling's \eqn{T^2} test.
}
