% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_mrfMCUSUM.R
\name{train_mrfMCUSUM}
\alias{train_mrfMCUSUM}
\title{Train mrfMCUSUM method}
\usage{
train_mrfMCUSUM(data, lags = 1, k = 0.9)
}
\arguments{
\item{data}{A multivariate time series in dataframe or matrix form.}

\item{lags}{The number of lags of each variable to be included in the design matrix.}

\item{k}{A tuning parameter for the MCUSUM, large k results in shorter memory.}
}
\value{
A named list including the plotting statistic, trained model, residuals, and constants.
}
\description{
Trains the mrfMCUSUM method.
}
