% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{gen_dat_lin}
\alias{gen_dat_lin}
\alias{gen_dat_ltl}
\alias{gen_dat_nlr}
\alias{gen_dat_ltm}
\title{Generate Sample Data}
\usage{
gen_dat_lin(n_ic = 1500, n_oc = 500, phi = 0.8, a_1 = 3, a_2 = 2)

gen_dat_ltl(
  n_ic = 1500,
  n_oc = 500,
  phi = 0.8,
  sin_scale = 1.5,
  a_1 = 3,
  a_2 = 2
)

gen_dat_nlr(n_ic = 1500, n_oc = 500, phi = 0.8, a_1 = 3, a_2 = 2)

gen_dat_ltm(
  n_ic = 1500,
  n_oc = 500,
  phi = 0.8,
  sin_scale = 1.5,
  a_1 = 3,
  a_2 = 2
)
}
\arguments{
\item{n_ic}{The number of observations which are in-control.}

\item{n_oc}{The number of observations which are out-of-control, i.e. faulty.}

\item{phi}{The strength of the autocorrelation in the base time series}

\item{a_1}{Increase in variance of x2 for fault 1.}

\item{a_2}{Increase in variance of x1, x2, and x3 for fault 3.}

\item{sin_scale}{The scale of the sine wave for non-stationarity.}
}
\value{
A list of four data frames where the first n_ic observations are
identical, but the last n_oc observations have experienced fault 1, 2, 3,
or no fault.
}
\description{
Generates a three dimensional multivariate time series for use in
simulation studies. Multivariate time series data are generated with
either linear or non-linear relationships between variables. Further,
the option to produce nonstationary data by adding a sine wave is included.
\n Each time a multivariate time series is generated four data frames are
produced. There are three different faults that are be introduced, and one
data frame in which no fault is introduced.
}
