% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_fd}
\alias{predict_fd}
\title{Apply Method}
\usage{
predict_fd(model, new_data, new_data_exog = NULL, pstat0 = NULL)
}
\arguments{
\item{model}{Output from the train_fd function.}

\item{new_data}{A multivariate time series in dataframe or matrix form.}
}
\value{
A named list including the plotting statistic and residuals.
}
\description{
Predicts the values of new data using the output from the train_fd
function. Returns the calculated residuals and the plotting statistic.
}
