% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train methods}
\usage{
train(
  data,
  method = "gruMCUSUM",
  lags = 1,
  k = 1.1,
  r = 0.3,
  center_scale = TRUE
)
}
\arguments{
\item{data}{A multivariate time series in dataframe or matrix form.}

\item{method}{An indicator of which model and fault detection method to use. Options are gruMCUSUM, mrfMCUSUM, varMCUSUM, varMEWMA, or htsquare.}

\item{lags}{The number of lags of each variable to be included in the design matrix.}

\item{k}{A tuning parameter for the MCUSUM, large k results in shorter memory.}

\item{r}{A tuning parameter for MEWMA, large r results in shorter memory.}

\item{center_scale}{A logical, whether or not to center and scale data before modeling.}
}
\value{
A named list including the plotting statistic, trained model, residuals, and constants.
}
\description{
Train the mrfMCUSUM, gruMCUSUM method. Also works for the resMCUSUM method
from Bodnar et al. (2017), which uses a VARMA(1, 1).
}
